+++ src/qt/client/client_controller.h
@@ -51,7 +51,7 @@
 typedef boost::function<void(const std::string&,
                              const boost::uint32_t&,
                              const boost::int16_t&,
-                             const float &)> IMNotifier;
+                             const double&)> IMNotifier;
 
 class ClientController : public QObject {
   Q_OBJECT
